``` feat(closed_loop_controller): 添加航向误差线速度自适应调节功能
添加了heading_error_threshold、heading_resume_threshold和
min_heading_speed_scale三个参数，实现根据航向误差自动调节线速度的功能。
当航向误差较小时保持完整线速度，误差在阈值间线性降低，误差很大时保留
最小线速度比例，避免原地旋转。同时增加了参数验证和日志输出功能。

fix(open_loop_controller): 修复轨迹ID类型转换问题

将轨迹ID从默认类型转换为long类型，解决可能的类型不匹配问题。
```

+++ src/planner/plan_manage_dmq/src/open_loop_controller.cpp
@@ -7,7 +7,7 @@ void trajectoryCallback(const scan_planner::BsplineConstPtr& message)
 {
   ROS_INFO("[open_loop_controller_dmq] diagnostic trajectory id=%d "
            "control_points=%lu knots=%lu",
-           message->traj_id,
+           static_cast<long>(message->traj_id),
            static_cast<unsigned long>(message->pos_pts.size()),
            static_cast<unsigned long>(message->knots.size()));
 }
